Redis-diff initial commit
Tool diffs two redis databases to see to what degree they match keys and
values.

Outputs delimited data that can be redirected into a logfile and
examined with other tools.

Code takes functions and inspiration from how redis-transfer was built
and builds on that basis. But it doesn't seem fitting to include this
tool in that repo.